+++ PKRenderer/Source/Core/RHI/RHI.h
@@ -13,7 +13,7 @@ namespace std
 
 namespace PKAssets
 {
-    enum class PKElementType : uint16_t;
+    enum class PKElementType : uint8_t;
     enum class PKDescriptorType : uint8_t;
     enum class PKShaderStage : uint8_t;  
     enum class PKShaderStageFlags : uint32_t;
